+++ src/passenger.cpp
@@ -10,7 +10,7 @@
 
 static uint32_t PID=0;
 
-passenger_t passenger_t::generate(int32_t from,uint64_t appear_time)
+passenger_t passenger_t::generate(int16_t from,uint64_t appear_time)
 {
 	using namespace constant;
 	int16_t dest;
@@ -18,13 +18,13 @@ passenger_t passenger_t::generate(int32_t from,uint64_t appear_time)
 	if(from<=base_floor)
 	{
 		dest=rand_between<int16_t>(base_floor+1,max_floor);
-		tolerance=static_cast<uint64_t>(((dest-from)*lift_up_tick+lift_up_first_extra_tick+lift_up_last_extra_tick)
+		tolerance=static_cast<uint64_t>(double((dest-from)*lift_up_tick+lift_up_first_extra_tick+lift_up_last_extra_tick)
 			*rand_between(tolerance_tick_rate_range)+0.5);
 	}
 	else
 	{
 		dest=rand_between<int>(0,2)?rand_between(min_floor,base_floor):base_floor;
-		tolerance=static_cast<uint64_t>(((from-dest)*lift_down_tick+lift_down_first_extra_tick+lift_down_last_extra_tick)
+		tolerance=static_cast<uint64_t>(double((from-dest)*lift_down_tick+lift_down_first_extra_tick+lift_down_last_extra_tick)
 			*rand_between(tolerance_tick_rate_range)+0.5);
 	}
 	return {from,dest,PID++,rand_between(passenger_weight_range),appear_time,tolerance,0};
